对代码进行重构，修复控制关键字状态定时任务出现的bug

+++ src/main/java/com/keymanager/monitoring/service/CustomerKeywordService.java
@@ -1132,9 +1132,11 @@ public class CustomerKeywordService extends ServiceImpl<CustomerKeywordDao, Cust
         List<Long> invalidCustomerKeywords = customerKeywordInvalidCountLogService.findInvalidCustomerKeyword();
         do {
             List<Long> subCustomerKeywordUuids = invalidCustomerKeywords.subList(0, (invalidCustomerKeywords.size() > 500) ? 500 : invalidCustomerKeywords.size());
-            customerKeywordDao.deleteBatchIds(subCustomerKeywordUuids);
-            logger.info("controlCustomerKeywordStatus:" + subCustomerKeywordUuids.toString());
-            invalidCustomerKeywords.removeAll(subCustomerKeywordUuids);
+            if (CollectionUtils.isNotEmpty(subCustomerKeywordUuids)) {
+                customerKeywordDao.deleteBatchIds(subCustomerKeywordUuids);
+                logger.info("controlCustomerKeywordStatus:" + subCustomerKeywordUuids.toString());
+                invalidCustomerKeywords.removeAll(subCustomerKeywordUuids);
+            }
         } while (invalidCustomerKeywords.size() > 0);
     }
